refactor: chande folder structure, minor variable fixes
