+++ src/features/PokemonList/CardList/CardList.css
@@ -10,7 +10,7 @@
   padding: 1rem;
   border-right: 1px solid #ccc;
   overflow-y: auto;
-  background-color: #f8f9fa;
+  /* background-color: #f8f9fa; */
   min-width: 30vw;
 }
 
@@ -27,7 +27,9 @@
 }
 
 .card-list-item {
-  margin-bottom: 2rem;
+  margin-bottom: 1rem;
+  display: flex;
+  align-items: center;
 }
 
 .card-list-button {
@@ -55,15 +57,15 @@
 }
 
 .card-list-item a {
-  margin: 0.25rem 0 0;
+  margin: 0 0.25rem;
   font-size: 0.9em;
-  color: #555;
+  /* color: #555; */
   width: 100%;
   text-align: left;
   padding: 0.75rem;
   border: 1px solid #ddd;
   border-radius: 8px;
-  background: white;
+  /* background: white; */
   cursor: pointer;
   transition: all 0.2s ease;
   border: 1px solid #e0e0e0;
