+++ src/features/PokemonList/PokemonDetails/PokemonDetails.css
@@ -2,7 +2,7 @@
   flex: 1;
   border-left: 2px solid #eee;
   padding: 2rem;
-  background-color: #f9f9f9;
+  /* background-color: #f9f9f9; */
   overflow-y: auto;
   min-width: 30vw;
 }
